requirements.txt
azure-devops
openai
requests
chevron
python-dotenv
pyyaml
regex
colorama
gitpython 